[Mac/Core] Enable macOS CPU test build path
- Wire macOS build targets to include A64 backend and mac-specific premake wiring
- Add macOS stubs for AudioMediaPlayer and FilePicker to avoid FFmpeg/UI blockers
- Add test-only force_mount_devkit cvar for CPU test binaries
- Adjust base/memory plumbing and kernel helpers needed for macOS CPU tests
- Align PPC test harness and opcode metadata for A64/arm64 test coverage
- Validation: user ran xenia-base-tests, xenia-cpu-ppc-tests, xenia-cpu-tests (arm64)

+++ src/xenia/cpu/breakpoint.cc
@@ -48,7 +48,8 @@ std::string Breakpoint::to_string() const {
     str += " " + functions[0]->name();
     return str;
   } else {
-    return std::string("x64 ") + xe::string_util::to_hex_string(host_address());
+    return std::string("x64 ") + xe::string_util::to_hex_string(
+                                     static_cast<uint64_t>(host_address()));
   }
 }
